# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar Payment API",
    "version": "0.1.0",
    "description": "API for creating and verifying Stellar network payments. Accept Stellar-based payments with simple links and status tracking.",
    "contact": {
      "name": "Stellar Payment API Support",
      "url": "https://github.com/emdevelopa/Stellar_Payment_API"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key for authenticating merchant requests. Obtain this key when registering a merchant."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/register-merchant": {
      "post": {
        "summary": "Register a new merchant",
        "tags": [
          "Merchants"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "business_name": {
                    "type": "string"
                  },
                  "notification_email": {
                    "type": "string",
                    "format": "email"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Optional free-form onboarding data (e.g. industry, country)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Merchant registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "merchant": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Merchant already exists"
          },
          "429": {
            "description": "Too many registration attempts"
          }
        }
      }
    },
    "/api/rotate-key": {
      "post": {
        "summary": "Rotate the authenticated merchant's API key",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New API key issued; the old key is immediately invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid x-api-key header"
          }
        }
      }
    },
    "/api/merchants/rotate-webhook-secret": {
      "post": {
        "summary": "Rotate the authenticated merchant's webhook signing secret",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grace_period_hours": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 168,
                    "description": "Optional override for old-secret grace period in hours (default 24)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New webhook secret issued; old secret remains valid until expiry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "webhook_secret": {
                      "type": "string"
                    },
                    "webhook_secret_old_expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "grace_period_hours": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/preview-receipt": {
      "post": {
        "summary": "Generate a preview HTML of the email receipt with custom branding",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HTML preview of the receipt",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/test-webhook": {
      "post": {
        "summary": "Send a test ping to a webhook URL",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/webhook-settings": {
      "get": {
        "summary": "Retrieve current webhook URL and masked webhook secret",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current webhook settings"
          }
        }
      },
      "put": {
        "summary": "Update the merchant's webhook endpoint URL",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhook_url": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook URL updated"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/regenerate-webhook-secret": {
      "post": {
        "summary": "Regenerate the merchant's webhook signing secret",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New webhook secret issued"
          },
          "401": {
            "description": "Missing or invalid x-api-key header"
          }
        }
      }
    },
    "/api/merchants/generate-api-key": {
      "post": {
        "summary": "Generate an API key using session authentication",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New API key issued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/merchants/rotate-api-key": {
      "post": {
        "summary": "Rotate API key with overlap period for seamless migration",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grace_period_hours": {
                    "type": "integer",
                    "description": "Hours the old key remains valid (0-168, default 24)",
                    "default": 24
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New API key generated with old key overlap period"
          }
        }
      }
    },
    "/api/merchants/api-key-status": {
      "get": {
        "summary": "Get current API key status and expiry information",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API key status information"
          }
        }
      }
    },
    "/api/merchants/set-api-key-expiry": {
      "put": {
        "summary": "Set an expiry date for the current API key",
        "tags": [
          "Merchants"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "expires_at"
                ],
                "properties": {
                  "expires_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "ISO 8601 datetime when the API key expires"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}